app.py 굳버전/requirements.txt
streamlit
pillow
streamlit-image-coordinates
